The android project
Connected the database to the android application. Project has
databasehelper class that checks for the database and a method to pull
the information from the database. Used query to pull name from Hero
table and displayed it through the main activity of the application. The
app will display all the heroes name.